misc/requirements.txt
bokeh==3.1.1
contourpy==1.1.0
gitdb==4.0.10
GitPython==3.1.31
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
smmap==5.0.0
tornado==6.3.2
tzdata==2023.3
xyzservices==2023.5.0
